Refactor: chair side
chair/chair_act.js: 足を上げる機構に関するコード
chair/measure.js: 重量計測に関するコード
chair/main.js: 機能を統合し, 通信を行うコード
