+++ chair/main.js
@@ -2,40 +2,14 @@ const sleep = msec => new Promise(resolve => setTimeout(resolve, msec));
 import nodeWebSocketLib from "websocket"; // https://www.npmjs.com/package/websocket
 import {RelayServer} from "./RelayServer.js";
 
-import { requestI2CAccess } from "./node_modules/node-web-i2c/index.js";
-import ADS1X15 from "@chirimen/ads1x15";
-
-import { init_motor, actuator_on, actuator_off } from "./chair_act.js"
+import { init_motor, actuator_on, actuator_off } from "./chair_act.js";
+import { init_measure, measure } from "./measure.js";
 
 var channel;
-var ads1115;
-
 var state = {
     threshold: 100,
     weight: 0,
 };
-
-async function init_all() {
-    const i2cAccess = await requestI2CAccess();
-    const port = i2cAccess.ports.get(1);
-    ads1115 = new ADS1X15(port, 0x48);
-    // If you uses ADS1115, you have to select "true", otherwise select "false".
-    await ads1115.init(true, 7); // High Gain
-    console.log("init complete");
-
-    await connect();
-    cast_threshold();
-
-    await init_motor();
-}
-
-async function connect(){
-    // webSocketリレーの初期化
-    var relay = RelayServer("chirimentest", "chirimenSocket" , nodeWebSocketLib, "https://chirimen.org");
-    channel = await relay.subscribe("isutribute_measure");
-    console.log("web socketリレーサービスに接続しました");
-    channel.onmessage = receiver;
-}
 
 function receiver(msg) {
     let data = msg.data;
@@ -46,9 +20,12 @@ function receiver(msg) {
     }
 }
 
-function change_threshold(data){
-    state.threshold = parseFloat(data.threshold);
-    console.log("new threshold: " + state.threshold );
+async function connect(){
+    // webSocketリレーの初期化
+    var relay = RelayServer("chirimentest", "chirimenSocket" , nodeWebSocketLib, "https://chirimen.org");
+    channel = await relay.subscribe("isutribute_measure");
+    console.log("web socketリレーサービスに接続しました");
+    channel.onmessage = receiver;
 }
 
 function cast_threshold() {
@@ -58,17 +35,28 @@ function cast_threshold() {
     });
 }
 
+function change_threshold(data){
+    state.threshold = parseFloat(data.threshold);
+    console.log("new threshold: " + state.threshold );
+}
+
+async function init() {
+    await connect();
+    cast_threshold();
+    await init_motor();
+    await init_measure();
+    console.log("init complete");
+}
+
 function judge(state) {
     return state.weight >= state.threshold;
 }
 
 async function main() {
     var firstTime = true;
-    const base = 22;
     let is_last_sit = false;
-    for (var i = 0; ;i++) {
-        var difA = await ads1115.read("0,1");  // p0-p1 differential mode
-        state.weight = difA - base;
+    for (;;) {
+        state.weight = await measure();
         console.log(state);
 
         if (judge(state)) {
@@ -77,8 +65,9 @@ async function main() {
                 "time": new Date().toISOString(),
             });
             console.log("sent");
-            if (!is_last_sit)
+            if (!is_last_sit) {
                 actuator_on();
+            }
             is_last_sit = true;
         } else if (is_last_sit) {
             channel.send({
@@ -88,10 +77,9 @@ async function main() {
             actuator_off();
             is_last_sit = false;
         }
-
         await sleep(1000);
     }
 }
 
-await init_all();
+await init();
 await main();
